requirements.txt
django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.10
drf-spectacular==0.27.2
django-cors-headers==4.5.0
pytest-django==4.9.0
